+++ Sources/UI/CreateRoomButton.swift
@@ -42,6 +42,11 @@ class CreateRoomButton: ButtonWithSpringAnimation {
         ])
 
         addTarget(self, action: #selector(didPress), for: [.touchUpInside])
+
+        layer.shadowOffset = CGSize(width: 0, height: 12)
+        layer.shadowColor = UIColor.black.withAlphaComponent(0.24).cgColor
+        layer.shadowOpacity = 1
+        layer.shadowRadius = 24
     }
 
     required init?(coder _: NSCoder) {